+++ key/aziot-key-client/src/lib.rs
@@ -49,15 +49,14 @@ impl Client {
     pub fn move_key_pair(&self, from: &str, to: &str) -> std::io::Result<()> {
         let mut stream = self.connector.connect()?;
 
-        let body = aziot_key_common_http::move_key_pair::Request {
+        let body = aziot_key_common_http::r#move::Request {
             from: from.to_owned(),
-            to: to.to_owned(),
         };
 
         request_no_content(
             &mut stream,
             &http::Method::POST,
-            format_args!("/keypair/move?api-version={}", self.api_version),
+            format_args!("/keypair/{}?api-version={}", to, self.api_version),
             Some(&body),
         )?;
