+++ key/aziot-key-common-http/src/lib.rs
@@ -179,11 +179,10 @@ pub mod delete {
     }
 }
 
-pub mod move_key_pair {
+pub mod r#move {
     #[derive(Debug, serde::Deserialize, serde::Serialize)]
     pub struct Request {
         pub from: String,
-        pub to: String,
     }
 }
